+++ Changelog.md
@@ -46,6 +46,7 @@ Bugfixes:
  * Commandline Interface: linking libraries with underscores in their name.
  * Type Checker: Fallback function cannot return data anymore.
  * Code Generator: Fix crash when sha3() was used on unsupported types.
+ * Code Generator: Manually set gas stipend for .send(0).
 
 Lots of changes to the documentation mainly by voluntary external contributors.